+++ src/main/java/de/samply/utils/ThrowingConsumer.java
@@ -1,5 +1,0 @@
-package de.samply.utils;
-
-public class ThrowingConsumer {
-
-}